Add files via upload
This function was made for the password game (https://neal.fun/password-game/). The game requires you to have a leap year and all numbers in the password to add up to 25. This function allows you to select the amount the digits should add to and will produce all leap years after the year 2000. I may update this to be from 1582(the start of the gregorian calendar), but I'm not sure yet.